+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ### **(V.7.11.5 Changes) (1.21.1 Minecraft)**
 
+#### Structures:
+
+Dying near the Essence Block in Sempiternal Sanctums will now move the player just outside the bounds of the sanctum's arena.
+This will allow Gravestone/Tombstone mods to continue spawning their block and not risk item duplication issues with arena nor break arena's invincible blocks.
+This does mean Bumblezone will no longer drop player items with unlimited despawn timer when Corail Tombstone is on and near Essence Block. Their tombstone block will spawn now.
+
 #### Mod Compat:
 
 Fixed Pollinated Pillar biome icon not showing properly in EMI Ores mod.